Added straight_drive_distance() and calibrate_straight_drive_distance for lego in drive.py, added PUSH_SENSOR in ports.py, made minor formatting changes, and replaced choose_to_calibrate() with a loop that has the same function but uses the new calibration function in init().
If get_digital_output(port) returns something other than the 1 or 0 value from the digital sensor, replace it with the function that does.

The straight_drive_distance_proportion (in straight_drive_distance()) has not been calibrated for any of the robots yet, and the current entries are a massive underestimate of the true value.